requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
psycopg2-binary
python-dotenv
razorpay
